教师端接口基本实现。

+++ src/main/java/SmartClass/Resource/CourseResource.java
@@ -1,20 +1,23 @@
-package SmartClass.Controller;
+package SmartClass.Resource;
 
-import SmartClass.Dao.AdminDao;
 import SmartClass.Dao.CourseDao;
+import SmartClass.Dao.StudentCourseDao;
 import SmartClass.Dao.TeacherDao;
-import SmartClass.DaoImp.AdminDaoImp;
 import SmartClass.DaoImp.CourseDaoImp;
+import SmartClass.DaoImp.StudentCourseDaoImp;
 import SmartClass.DaoImp.TeacherDaoImp;
-import SmartClass.POJO.Administrator;
 import SmartClass.POJO.Course;
 import SmartClass.POJO.Teacher;
 import SmartClass.dbutil.DbUtil;
+import SmartClass.tool.HttpSessionUtil;
 import org.json.JSONArray;
 import org.json.JSONObject;
 
+import javax.servlet.http.HttpServletRequest;
 import javax.ws.rs.*;
+import javax.ws.rs.core.Context;
 import javax.ws.rs.core.MediaType;
+import java.sql.Date;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -26,15 +29,22 @@ import java.util.Map;
 public class CourseResource
 {
     private final static String CHARSET_UTF_8 = "charset=utf-8";
-
     private CourseDao courseDao = new CourseDaoImp();
     private TeacherDao teacherDao = new TeacherDaoImp();
+    private StudentCourseDao studentCourseDao = new StudentCourseDaoImp();
     /*添加课程*/
     @POST
     @Produces(MediaType.APPLICATION_JSON)
-    public Map<String,Object> addCourse(String reqText)
+    public Map<String,Object> addCourse(String reqText, @Context HttpServletRequest request)
     {
         Map reply = new HashMap<String,Object>();
+        /*管理员是否登录*/
+        if (!HttpSessionUtil.islogin(request,"role","admin"))
+        {
+            reply.put("status", 1000);
+            reply.put("msg","此操作是只能由管理员执行，请先登录！");
+            return reply;
+        }
         /*处理请求*/
         JSONObject jsReq = new JSONObject(reqText);
         String courseName = jsReq.getString("courseName");
@@ -47,6 +57,8 @@ public class CourseResource
             course.setCourseName(courseName);
             course.setTeacherByTeacherId(teacher);
             course.setCheckinFlag((byte)0);
+            course.setCheckinTime((short)0);
+            course.setCheckinDate(new Date(2000,1,1));
             course.setAnswerFlag((byte)0);
             course.setTimeCreated(DbUtil.now());
             course.setTimeModified(DbUtil.now());
@@ -68,9 +80,17 @@ public class CourseResource
     @Path("{courseId}")
     @PUT
     @Produces(MediaType.APPLICATION_JSON + ";" + CHARSET_UTF_8)
-    public Map<String,Object> updataCourse(@PathParam("courseId")short courseId, String reqText)
+    public Map<String,Object> updataCourse(@PathParam("courseId")short courseId, String reqText,
+                                           @Context HttpServletRequest request)
     {
         Map reply = new HashMap<String,Object>();
+        /*管理员是否登录*/
+        if (!HttpSessionUtil.islogin(request,"role","admin"))
+        {
+            reply.put("status", 1000);
+            reply.put("msg","此操作是只能由管理员执行，请先登录！");
+            return reply;
+        }
         /*处理请求*/
         JSONObject jsReq = new JSONObject(reqText);
         String courseName = jsReq.getString("courseName");
@@ -102,13 +122,25 @@ public class CourseResource
     @Path("{courseId}")
     @DELETE
     @Produces(MediaType.APPLICATION_JSON + ";" + CHARSET_UTF_8)
-    public Map<String,Object> deleteCourse(@PathParam("courseId") int courseId)
+    public Map<String,Object> deleteCourse(@PathParam("courseId") short courseId,
+                                           @Context HttpServletRequest request)
     {
         Map reply = new HashMap<String,Object>();
-
+        /*管理员是否登录*/
+        if (!HttpSessionUtil.islogin(request,"role","admin"))
+        {
+            reply.put("status", 1000);
+            reply.put("msg","此操作是只能由管理员执行，请先登录！");
+            return reply;
+        }
         try
         {
-            /*删除课程*/
+            /*删除课程和学生的关系*/
+            studentCourseDao.deleteByCourseId(courseId);
+            /*删除课程,由于级联操作，
+            *会同时删除sgininfo表中和courseId有关的记录
+            *会同时删除comminication表中和courseId有关的记录
+            * */
             courseDao.deleteById((short)courseId);
         }catch (Exception e)
         {
@@ -126,9 +158,16 @@ public class CourseResource
     /*得到所有课程*/
     @GET
     @Produces(MediaType.APPLICATION_JSON + ";" + CHARSET_UTF_8)
-    public String getAllCourse()
+    public String getAllCourse(@Context HttpServletRequest request)
     {
         JSONObject reply = new JSONObject();
+        /*管理员是否登录*/
+        if (!HttpSessionUtil.islogin(request,"role","admin"))
+        {
+            reply.put("status", 1000);
+            reply.put("msg","此操作是只能由管理员执行，请先登录！");
+            return reply.toString();
+        }
         try
         {
             /*查询*/
